Move lazy-loaded Thanks component to containers
